pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openkb"
version = "0.1.4"
description = "OpenKB: Open LLM Knowledge Base, powered by PageIndex (subprocess executor fork)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Kylin", email = "quanqi@pageindex.ai"},
    {name = "Ray", email = "ray@vectify.ai"},
    {name = "Mizelan", email = "mizelan@users.noreply.github.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "rag", "retrieval", "knowledge-base", "llm", "pageindex", "agents", "document"]
dependencies = [
    "pageindex==0.3.0.dev1",
    "markitdown[all]",
    "click>=8.0",
    "watchdog>=3.0",
    "json-repair",
    "pyyaml",
    "python-dotenv",
    "pymupdf",
    "prompt_toolkit>=3.0",
    "rich>=13.0",
    "networkx>=3.0",
    "python-louvain>=0.16",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.urls]
Repository = "https://github.com/VectifyAI/OpenKB"
Homepage = "https://github.com/VectifyAI/OpenKB"
Issues = "https://github.com/VectifyAI/OpenKB/issues"

[project.scripts]
openkb = "openkb.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["openkb"]
